* read correct angles in Nexus files where angles are saved in radians (cctbx#192)
* fix file processing issues where % characters are in the path (cctbx#214)
* allow reading SMV ADSC format files with non-integer pedestal (cctbx#216)

+++ format/FormatSMVRigakuSaturnNoTS.py
@@ -157,7 +157,7 @@ class FormatSMVRigakuSaturnNoTS(FormatSMVRigaku):
             pixel_size,
             image_size,
             (underload, overload),
-            pedestal=int(self._header_dictionary.get("IMAGE_PEDESTAL"), 0),
+            pedestal=float(self._header_dictionary.get("IMAGE_PEDESTAL"), 0),
         )
 
     def _beam(self):
